+++ packages/zwave-js/src/lib/driver/Driver.ts
@@ -15,8 +15,10 @@ import {
 	deserializeCacheValue,
 	Duration,
 	highResTimestamp,
+	LogConfig,
 	SecurityManager,
 	serializeCacheValue,
+	updateLogConfig,
 	ValueMetadata,
 	ZWaveError,
 	ZWaveErrorCodes,
@@ -92,11 +94,7 @@ import {
 import { getDefaultPriority, Message } from "../message/Message";
 import { isNodeQuery } from "../node/INodeQuery";
 import type { ZWaveNode } from "../node/Node";
-import {
-	InterviewStage,
-	NodeInterviewFailedEventArgs,
-	NodeStatus,
-} from "../node/Types";
+import { InterviewStage, NodeStatus } from "../node/Types";
 import type { FileSystem } from "./FileSystem";
 import {
 	createSendThreadMachine,
@@ -150,6 +148,11 @@ export interface ZWaveOptions {
 		 */
 		nodeInterview: number; // [1...10], default: 5
 	};
+
+	/**
+	 * Optional log configuration
+	 */
+	logConfig?: LogConfig;
 
 	/**
 	 * @internal
@@ -455,6 +458,11 @@ export class Driver extends EventEmitter {
 		) as ZWaveOptions;
 		// And make sure they contain valid values
 		checkOptions(this.options);
+
+		if (this.options.logConfig) {
+			updateLogConfig(this.options.logConfig);
+		}
+
 		this.cacheDir = this.options.cacheDir;
 
 		// register some cleanup handlers in case the program doesn't get closed cleanly
@@ -798,15 +806,6 @@ export class Driver extends EventEmitter {
 		}
 	}
 
-	private emitNodeInterviewFailed(
-		node: ZWaveNode,
-		args: NodeInterviewFailedEventArgs,
-	): void {
-		// For compatibility, we emit the arguments as the third parameter until the next major version
-		// TODO: remove this compatibility layer and emit the args as the 2nd parameter directly
-		node.emit("interview failed", node, args.errorMessage, args);
-	}
-
 	private retryNodeInterviewTimeouts = new Map<number, NodeJS.Timeout>();
 	/**
 	 * @internal
@@ -841,7 +840,7 @@ export class Driver extends EventEmitter {
 						`Interview attempt (${node.interviewAttempts}/${maxInterviewAttempts}) failed, node is dead.`,
 						"warn",
 					);
-					this.emitNodeInterviewFailed(node, {
+					node.emit("interview failed", node, {
 						errorMessage: "The node is dead",
 						isFinal: true,
 					});
@@ -856,7 +855,7 @@ export class Driver extends EventEmitter {
 						`Interview attempt ${node.interviewAttempts}/${maxInterviewAttempts} failed, retrying in ${retryTimeout} ms...`,
 						"warn",
 					);
-					this.emitNodeInterviewFailed(node, {
+					node.emit("interview failed", node, {
 						errorMessage: `Attempt ${node.interviewAttempts}/${maxInterviewAttempts} failed`,
 						isFinal: false,
 						attempt: node.interviewAttempts,
@@ -876,7 +875,7 @@ export class Driver extends EventEmitter {
 						`Failed all interview attempts, giving up.`,
 						"warn",
 					);
-					this.emitNodeInterviewFailed(node, {
+					node.emit("interview failed", node, {
 						errorMessage: `Maximum interview attempts reached`,
 						isFinal: true,
 						attempt: maxInterviewAttempts,